# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"InternalServerError": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"error": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"error",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessResponseCreateLoanRequest": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoanRequestDocument": {
				"properties": {
					"nombre": {
						"type": "string"
					},
					"enlace": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"enlace"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoanRequest": {
				"properties": {
					"cliente_id": {
						"type": "number",
						"format": "double"
					},
					"tipo_prestamo_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"monto_solicitado": {
						"type": "number",
						"format": "double"
					},
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"documentos": {
						"items": {
							"$ref": "#/components/schemas/LoanRequestDocument"
						},
						"type": "array"
					}
				},
				"required": [
					"cliente_id",
					"tipo_prestamo_id",
					"empresa_id",
					"monto_solicitado",
					"emp_id",
					"documentos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LoanRequest_": {
				"properties": {
					"cliente_id": {
						"type": "number",
						"format": "double"
					},
					"tipo_prestamo_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"monto_solicitado": {
						"type": "number",
						"format": "double"
					},
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"documentos": {
						"items": {
							"$ref": "#/components/schemas/LoanRequestDocument"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "finantek_api",
		"version": "0.0.1",
		"description": "API fir finantek",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Daniel Valdez | hello@daniel-valdez.com | https://daniel-valdez.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/solicitudes-prestamo/registrar": {
			"post": {
				"operationId": "CreateLoanRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponseCreateLoanRequest"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoanRequest"
							}
						}
					}
				}
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}": {
			"get": {
				"operationId": "GetAllLoanRequestsByCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"items": {
												"$ref": "#/components/schemas/LoanRequest"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{id}": {
			"get": {
				"operationId": "GetLoanRequestByIdAndCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/LoanRequest"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{id}/editar": {
			"put": {
				"operationId": "UpdateLoanRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LoanRequest_"
							}
						}
					}
				}
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{id}/eliminar": {
			"delete": {
				"operationId": "DeleteLoanRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{solicitud_id}/actualizar-estado": {
			"put": {
				"operationId": "UpdateLoanRequestStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "solicitud_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"empleado_id": {
										"type": "number",
										"format": "double"
									},
									"mensaje": {
										"type": "string"
									},
									"nuevo_estado": {
										"type": "string"
									}
								},
								"required": [
									"empleado_id",
									"mensaje",
									"nuevo_estado"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}